+++ app/src/main/java/com/levent_j/potato/utils/Util.java
@@ -4,7 +4,7 @@ package com.levent_j.potato.utils;
  * Created by levent_j on 16-5-20.
  */
 public class Util {
-    public static int Minute2Second(int minute){
+    public static double Minute2Second(double minute){
         //*60000
         return minute*5000;
     }
